Added a functionality to store new requests
When a post request arrives, the request handler function is called and two variables are created "newId" & "newTour". The "newId" is created by counting total numbers of elements and generating a new number. "newTour" is obtained by merging the object with the "newId" with the request object. Finally "newTour" is added with the exisiting elements and then values are saved in the json file and then a response is sent to the client containing the "newTour".