Added a route for one tour with GET method
The route will listen to requests coming on "/api/v1/tours/:id" with GET method and then respond to the request by calling a request handler function. In this function body the "id" variable is created by storing the variable from the url and then "id" is checked with "tours" object. According to the match either a 200 or 400 status code is sent to the client.